requirements.txt
fastapi
httpx
uvicorn
supabase
python-dotenv
